// .worktrees/contract-analytics-filters/src/types.ts
export interface File {
  name:       string;
  url:        string;
  type:       string;
  size:       string;
  _id:        string;
  uploadedAt: string;
}